fixes rollup internal key
previously, rollup internal keys would use the time entry full
description.
however, when inserted into jira worklog, only yhe issue ID would be
part of the key (along with the day date).
as a result, multiple entries for the same issue ID but with different
descriptions created multiple rollups for a single day, and only the
first rollup was inserted properly in jira, and the others were rejected
by toji because the rollup key was already present in jira.

this commit fixes the internal rollup key and aggregates time based on
Jira issue ID, not toggle entry description